# back-end/laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mentes notáveis",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/addresses/{id}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get a specific address",
                "operationId": "f5f8e416d457384bd0b870fd03cf490b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}